+++ run.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+# Start container
+docker compose up -d