+++ stop.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+# Stop container
+docker compose down